requirements.txt
Flask==3.0.3
msal==1.30.0
pymongo==4.8.0
python-dotenv==1.0.1
certifi==2024.2.2
gunicorn==23.0.0